Move client async function support into SmokeAWSHttp package.
