+++ Package.swift
@@ -126,12 +126,12 @@ let package = Package(
             targets: ["SmokeAWSMetrics"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-nio.git", from: "2.0.0"),
-        .package(url: "https://github.com/apple/swift-nio-ssl.git", from: "2.0.0"),
+        .package(url: "https://github.com/apple/swift-nio.git", from: "2.33.0"),
+        .package(url: "https://github.com/apple/swift-nio-ssl.git", from: "2.14.0"),
         .package(url: "https://github.com/apple/swift-log", from: "1.0.0"),
         .package(url: "https://github.com/apple/swift-metrics.git", "1.0.0"..<"3.0.0"),
         .package(url: "https://github.com/LiveUI/XMLCoding.git", from: "0.4.1"),
-        .package(url: "https://github.com/amzn/smoke-http.git", from: "2.8.5"),
+        .package(url: "https://github.com/amzn/smoke-http.git", from: "2.9.0"),
         .package(url: "https://github.com/apple/swift-crypto.git", from: "1.0.0"),
     ],
     targets: [
@@ -292,7 +292,6 @@ let package = Package(
         .target(
             name: "_SmokeAWSHttpConcurrency", dependencies: [
                 .target(name: "SmokeAWSHttp"),
-                .product(name: "_SmokeHTTPClientConcurrency", package: "smoke-http"),
             ]),
         .target(
             name: "SmokeAWSMetrics", dependencies: [
