+++ Sources/SmokeAWSHttp/AWSQueryClientProtocol+asyncSupport.swift
@@ -1,14 +1,26 @@
+// Copyright 2018-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
-//  AWSClientProtocol.swift
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+//  AWSQueryClientProtocol+asyncSupport.swift
+//  SmokeAWSHttp
 //
 
-#if compiler(>=5.5)
+#if compiler(>=5.5) && canImport(_Concurrency)
 
 import SmokeHTTPClient
 import SmokeAWSCore
 import NIO
 import SmokeAWSHttp
-import _SmokeHTTPClientConcurrency
 
 public extension AWSQueryClientProtocol {
     @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
